requirements.txt
pyTelegramBotAPI
python-dotenv
requests
openclaw
tenacity
anthropic
